pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amharic-sentiment"
version = "1.0.0"
description = "Deep learning-based sentiment analysis for Amharic text"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Fikerte Shalemo"}
]
keywords = [
    "sentiment-analysis",
    "amharic",
    "nlp",
    "deep-learning",
    "tensorflow",
    "ethiopian-languages"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "numpy>=1.21.0,<2.0.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "tensorflow>=2.10.0,<3.0.0",
    "gensim>=4.0.0",
    "nltk>=3.6.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "pyyaml>=5.4.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
]
notebook = [
    "jupyter>=1.0.0",
    "ipywidgets>=7.7.0",
]

[project.scripts]
amharic-sentiment = "amharic_sentiment.cli:main"

[project.urls]
Homepage = "https://github.com/Fikerdeep/Amharic-Sentiment-Analysis"
Repository = "https://github.com/Fikerdeep/Amharic-Sentiment-Analysis"
Issues = "https://github.com/Fikerdeep/Amharic-Sentiment-Analysis/issues"

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*", "notebooks", "notebooks.*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["amharic_sentiment"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=amharic_sentiment --cov-report=term-missing"
